feat: 新增 showPageSize、showPageNumber API
