+++ src/pagination/pagination.md
@@ -17,6 +17,8 @@ defaultPageSize | Number | 10 | 分页总页数。非受控属性 | N
 pageSizeOptions | Array | () => [5, 10, 20, 50] | 分页大小控制器，值为 [] 则不显示。TS 类型：`Array<number | { label: string; value: number }>` | N
 showFirstAndLastPageBtn | Boolean | false | 是否显示跳转首页尾页页码控制器 | N
 showJumper | Boolean | false | 是否显示跳转页码控制器 | N
+showPageNumber | Boolean | true | 是否显示页码控制器 | N
+showPageSize | Boolean | true | 是否显示分页数量控制器 | N
 showPreviousAndNextBtn | Boolean | true | 是否显示跳转前后页页码控制器 | N
 size | String | medium | 分页组件尺寸。可选项：small/medium | N
 theme | String | default | 分页组件风格。可选项：default/simple | N
