+++ src/pagination/usage/props.json
@@ -18,6 +18,18 @@
     "options": []
   },
   {
+    "name": "showPageNumber",
+    "type": "Boolean",
+    "defaultValue": true,
+    "options": []
+  },
+  {
+    "name": "showPageSize",
+    "type": "Boolean",
+    "defaultValue": true,
+    "options": []
+  },
+  {
     "name": "showPreviousAndNextBtn",
     "type": "Boolean",
     "defaultValue": true,
